requirements.txt
Flask
Flask-WTF
Flask-Bcrypt
Flask-Login
Flask-Migrate
mysql-connector-python
bcrypt
